backend/requirements.txt
asgiref==3.7.2
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
mysqlclient==2.2.4
python-decouple==3.8
pandas==2.2.2
xlrd >= 2.0.1
django-filter==24.2
